Fixing mundane potion meta.
The metadata of mundane potions is 64.

To calculate the metadata of the splash mundane potions you use this formula `64 >> 32 | (0 << 5) | (0 << 6) | (0 << 13) | (1 << 14)`. This is the same formula used to calculate the metadata for splash awkward potions, and splash thick potions.

See more information here https://minecraft.gamepedia.com/Potion#Unbrewable_potions

I have also removed the extended mundane potions as they do not exist in the game, and I have no idea where the id values listed for them in this file originated.