Ensure LOGGER is not equal to plugin logger before setting parent
PaperMC/Paper#890 introduced minor changes to the plugin logger to
make plugin prefixes independent from the logging framework. As a
result, the plugin logger now uses the plugin name as name.

Essentials uses the same logger name and redirects messages to the
plugin logger by setting it as a parent. However, in newer Paper
builds both loggers are now equal, resulting in infinite loops when
logging messages because the logger has itself as parent.

Make sure the two loggers are not equal before setting the parent to
avoid the following warning on newer Paper builds:

    [Essentials] Ignoring attempt to change parent of plugin logger