javadoc metaspec 1st commit This commit adds javadoc meta tags to every class in the project
Please NOTE:  Future commits will be required to edit the original javadoc contributors name (LoopyD) out of the doucmentation's @author tags.  javadoc-fix was able to provide some, but not all of this information, and so inserted the default.  Please insert yourself as Author where prompted in the source code in future commits to this fork.

This commit passes javadoc metaspec unit tests and will allow the project to have documentation generated for it.  It reports 0 errors and 0 warnings.