Add "localNoOne" text.
I would like to suggest the addition of text to the "localNoOne" entry. This message is supposed to be sent when a player sends a chat message with no players nearby if the chat radius feature is enabled. Currently, this does not occur due to the fact that there is no message defined in "messages.properties". Adding this message would be helpful to players since the plugin would notify them if nobody was able to hear their chat message. I would also suggest the addition of a configuration option which controls whether or not this message is sent so that a server administrator can turn off the feature if they don't want it enabled.