send-info-after-death
A possibly fixed version of https://github.com/EssentialsX/Essentials/pull/2912. I'm not very skilled in coding but here's my best shot on making the suggested changes. I couldn't have done this without seeing darbyjack's PR. I'm not sure if this is the standard way of modifying a PR, sorry if it's not.