Amended my previous commit. Before, I introduced a method to multiply a matrix by a vector, even though there were only ever 3 matrices that would ever be inputted. Instead, I took those three cases and simplified the math (it actually really simplifies) and hard coded it in. If there is ever a need for a method to multiply a matrix by a vector it wouldn't even belong in this class anyway.
Side note, the only thing that changed functionally in this commit is that the random point is no longer rounded to an integer, so the offset may be something like 5.83. I have no idea if that matters, but it can be changed very easily, and I'll allow whoever pulls this to make that decision.