Make currency symbol standardized in Kit
Kits should be in a standardized format.
Having monetary rewards in kits should not break based off of a config value oriented around currency display (suffix).
Additionally, the dollar sign should be the standard money symbol which works on all servers and should not be tied to the server's individual currency symbol.
Note that the server's individual currency symbol will still work but probably should not be used due its volatility.