Always update usernames on join
Always call ModernUserMap#getUser(Player) during join to ensure the
fetched/created User object has the most up-to-date name on the underlying
Player object. Additionally, ensure that ModernUUIDCache#updateCache is always
called during #getUser and not just when we generate a new User object (which
always never happens by the time the call is made :skull:).