Fix teleport permission bypass exploit
Players can bypass the "tpaccept" permission check when teleport delay is enabled.
This can be done by an exploiter getting the target to /tpa to them, then, the exploiter starts a teleport timer to a location where they do not have "tpaccept" and then typing /tpaccept before that timer finishes.
This will cause the first timer to teleport the exploiter to the location where they do not have "tpaccept" and then the timer for the target player will finish, resulting in the target teleporting to the exploiters location.