Fix NPE while syncing offline users for discord line module
Only would happen for users who haven't joined
(or been loaded by different parts of the plugin)
since the last restart. This change first of all switches to
a method to fetch users which will update the User base to the
UUIDPlayer dummy base. Secondly, this change will not update
the base of a User to a UUIDPlayer dummy unless the base is
currently null (which would be the case in the condition
described above).