Fix startup issue with potion providers
The LegacyPotionMetaProvider was initializing a map with enums that don't exist on modern versions. Since this was a static variable, this map got initialized when the test method was called, leading to a NoSuchFieldError.

Also improved resilience of these provider-type issues again by catching a Throwable rather than an Exception in the ProviderFactory.